+++ packages/ApiTests/src/dapi/packages.ts
@@ -27,4 +27,4 @@ category('Dapi: packages', () => {
   test('readCsv error', async () => {
     await expectExceptionAsync(() => _package.files.readCsv('datasets/noFile.csv').then());
   });
-});
+}, {node: true});