+++ packages/ApiTests/src/dataframe/dataframe.ts
@@ -156,7 +156,7 @@ category('DataFrame: Methods', () => {
     expect(t2.columns.byName('x').getNumber(0) === 9);
     expect(t2.columns.byName('z').getNumber(1) === 7);
     expect(t2.columns.byName('y').getNumber(2) === 6);
-  }, {skipReason: typeof process !== 'undefined' ? 'NodeJS environment' : undefined});
+  }, {node: false, skipReason: typeof process !== 'undefined' ? 'NodeJS environment' : undefined});
 
   test('toJson | fromJson', async () => {
     const t = createDf();
@@ -218,7 +218,7 @@ Canada,4`;
     expect(df.columns.length, 2);
     expect(df.rowCount, 4);
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});
 
 // Column
 
@@ -463,7 +463,7 @@ category('DataFrame: Column', () => {
     col.init((_: any) => 'val');
     expect(col.get(0), 'val');
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});
 
 // ColumnList
 
@@ -670,7 +670,7 @@ category('DataFrame: ColumnList', () => {
     const df1 = createDf();
     df1.columns.toString();
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});
 
 // Row
 
@@ -685,7 +685,7 @@ category('DataFrame: Row', () => {
   test('toDart', async () => {
     expect(typeof row.toDart(), 'object');
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});
 
 // RowList
 
@@ -786,4 +786,4 @@ category('DataFrame: RowList', () => {
     const str: string = createDf().rows.toString();
     expect(str.startsWith('(Instance of '), true);
   });
-}, {owner: 'aparamonov@datagrok.ai'});
+}, {owner: 'aparamonov@datagrok.ai', node: true});